Replace raw.github.com by rawgithub.com in README
The reason are well explained in rawgithub.com, but basically github sends (on purpose) the wrong headers, so resources can be wrongly/not interpreted by browsers. I'm not affiliated with rawgithub.com (which is a free service anyway) and I know it could die anytime, but it's still better that the GH raw thing which is good for downloading only.